CB-402 refactor: modify Daily main page
데일리 기록 메인 페이지 리팩토링
- 이름변경
- 불필요한 기능 삭제
