--- src/pages/Daily/Feeds/index.tsx
@@ -1,9 +1,0 @@
-import Layout from '@/components/layout/Layout';
-
-export default function DailyFeeds() {
-  return (
-    <Layout header title="급여량 기록" canGoBack>
-      <div>급여량 기록 페이지</div>
-    </Layout>
-  );
-}
